ml-engine/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
scipy==1.11.3
lightgbm==4.0.0

# Deep Learning
torch==2.0.1
transformers==4.33.3
tensorflow==2.13.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23

# Data Processing
pydantic==2.4.2

# Utilities
python-multipart==0.0.6
pytest==7.4.3
jupyter==1.0.0
